Replace template tag with include
The logic in the template tag has been moved to the views file.
