+++ cbv/templates/cbv/klass_detail.html
@@ -35,7 +35,7 @@
 
 
 {% block nav %}
-    {% nav projectversion klass.module klass %}
+    {% include "cbv/includes/nav.html" %}
 {% endblock nav %}
