+++ cbv/templates/cbv/module_detail.html
@@ -1,12 +1,11 @@
 {% extends 'base.html' %}
-{% load cbv_tags %}
 
 
 {% block title %}{{ module }}{% endblock %}
 
 
 {% block nav %}
-    {% nav projectversion module %}
+    {% include "cbv/includes/nav.html" %}
 {% endblock nav %}
 
 {% block page_header %}<h1>{{ module }}</h1>{% endblock %}
