+++ cbv/templates/home.html
@@ -1,10 +1,9 @@
 {% extends 'base.html' %}
 
-{% load cbv_tags %}
 {% load static %}
 
 {% block nav %}
-    {% nav projectversion %}
+    {% include "cbv/includes/nav.html" %}
 {% endblock %}
